add dev more to build.sh that doesnt delete the build dir

+++ build.sh
@@ -7,4 +7,9 @@ cp -r src build/chrome
 node mixin.js firefox/manifest.json build/firefox/manifest.json
 cd build/chrome; zip -r ../../dist/chrome.zip *; cd ../..
 cd build/firefox; zip -r ../../dist/firefox.xpi *; cd ../..
-rm -rf build
+
+if [ -z "$1" ]
+  then
+    # No first argument supplied, deleting build dir
+	rm -rf build
+fi
